fix: address copilot review comments
- Use scheduler_config.subnet_messages_per_round_instruction_limit in
  test_utilities drain_subnet_messages instead of recomputing the fraction.
- Fix i32 overflow in DTS test WAT loop: 3_000_000_000 exceeds i32::MAX,
  causing the loop to exit immediately. Switch to i64 counter/compare.

+++ rs/state_machine_tests/tests/flexible_ordering.rs
@@ -468,10 +468,10 @@ fn test_dts_execution_completes() {
             r#"(module
                 (import "ic0" "msg_reply" (func $msg_reply))
                 (func $run
-                    (local $i i32)
+                    (local $i i64)
                     (loop $loop
-                        (local.set $i (i32.add (local.get $i) (i32.const 1)))
-                        (br_if $loop (i32.lt_s (local.get $i) (i32.const 3000000000)))
+                        (local.set $i (i64.add (local.get $i) (i64.const 1)))
+                        (br_if $loop (i64.lt_s (local.get $i) (i64.const 3000000000)))
                     )
                     (call $msg_reply))
                 (memory $memory 1)
